+++ src/components/Card.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 
 const Card = (vehiculo) => {
   return (
-    <div>
+    <div className='col-3'>
       <p>Patente: {vehiculo.patente}</p>
     </div>
   )
